Adapt naming and size
